requirements_no_deps.txt
# Documentation: https://kafka-python.readthedocs.io/en/master/index.html
# Package: #https://pypi.org/project/kafka-python/
kafka-python==2.0.2